requirements.txt
pyTelegramBotAPI==4.14.0
flask==2.3.3
python-dotenv==1.0.0